refactor(feature-020): discharge the activation precondition on fixture doors
Move the T047 dark-runtime oracle suite in-crate as
runtime.rs::dark_runtime_oracles and the T048 wrapper contract-shape
oracle as public_api.rs::dark_wrapper_oracles, both verbatim with only
import paths changed; delete tests/runtime_dark_v11.rs and trim the
external delta file to its remaining export-delta oracle. With the
oracles in-crate, every fixture door in runtime.rs, embedded.rs, and
public_api.rs tightens from any(test, feature server) to
all(test, feature server) and stops shipping in the release binary,
exactly as the recorded precondition at the for_test_root doc demanded.
Private items whose only remaining consumers are the in-crate oracles
carry not(test)-scoped dead_code allows until the T066 wiring gives
them production callers. Seal pins refreshed from the Rust oracle's
printed actuals after rustfmt; embed feature gate observed green.
